refactor: move decision, project, taskrun boundary+control to domain modules (Phase 2)
- decision-domain: move DecisionInbox (control), DecisionsView (boundary), DecisionSpec (test)
- project-domain: move ProjectStorageService (control), ProjectModelsSpec (test)
- taskrun-domain: move TasksView, ReportsView, GraphView, LogsView (boundary), TaskRunFoldSpec (test)
- config-domain: move ConfigRepository trait (unblocks decision-domain dep on ConfigRepository)
- Update HtmlViews, ApplicationDI, controllers to use new module locations
